% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place-package.R
\docType{data}
\name{city}
\alias{city}
\title{city}
\description{
Data frame of city
}
\details{
'name', 'lon', 'lat', 'country', 'pop', 'capital'

Copied from maps package 'world.cities' with some renaming.
}
