% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent_grid.R
\name{extent_grid}
\alias{extent_grid}
\title{Extent grid}
\usage{
extent_grid(extent, dim = NULL, projection = NULL, lonlat = FALSE)
}
\arguments{
\item{extent}{numeric vector, xmin,xmax,ymin,ymax}

\item{dim}{optional dimension, 1 number or 2}

\item{projection}{optional projection string to store}

\item{lonlat}{if set uses cosine latitude scaling for width for longitude}
}
\value{
list of extent, dimension, projection
}
\description{
Helper function to create a grid dimension with a sensible aspect ratio.
}
\examples{
extent_grid(c(0, 1, 0, 2))
}
