% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where.R
\name{where}
\alias{where}
\alias{print.where}
\title{Where}
\usage{
where(x, y = NULL, wh = 1852 * 60)

\method{print}{where}(x, ...)
}
\arguments{
\item{x}{longitude}

\item{y}{latitude}

\item{wh}{width and height (can be a single value) assumed to be metres}

\item{...}{in the print method}
}
\value{
a where structure, the location and width/height around it
}
\description{
Specify a region by location point (longitude, latitude) and a width and height
around that point.
}
\details{
Width and height will be assumed to be in metres usually, but no units are recorded.

If no input point is given one is generated randomly. The default wh value is approximately the distance
along one degree of a great circle on Earth.
}
\examples{
where()

where(147, -42, wh = c(2000, 3000))
}
