% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place-package.R
\docType{package}
\name{place-package}
\alias{place-package}
\alias{_PACKAGE}
\title{place: Captures an Intended Region Specified as a Point and a Distance}
\description{
Specify a region by location and distance either side.This package doesn't do anything with this information, it just provides a way to capture this one idea.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdsumner/where}
  \item Report bugs at \url{https://github.com/mdsumner/where/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
